requirements.txt
# FastAPI - 现代、快速的Web框架
fastapi==0.104.1

# Uvicorn - ASGI服务器，用于运行FastAPI应用
uvicorn[standard]==0.24.0

# Motor - 异步MongoDB驱动
motor==3.7.1

# Pydantic - 数据验证和序列化
pydantic>=2.7.0

# Python-multipart - 处理文件上传
python-multipart==0.0.6

# docx2pdf - DOCX转PDF转换
docx2pdf==0.1.8

# python-docx - 处理DOCX文件
python-docx==1.1.0

# aiofiles - 异步文件操作
aiofiles==23.2.1

# python-dotenv - 环境变量管理
python-dotenv==1.0.0

# loguru - 现代化的日志记录
loguru==0.7.2 